Modify project structure to connect to MongoDB
This commit adds a DbSettings class (with an interface) to communicate
with `appsettings.json` file, where the database settings (connection
string, name of db and collection) will be saved. This class is added as
a singleton in Startup class, along with the already added singleton of
Service class.

The service class is completely overhauled to make it work with the new
Settings class, which is injected into the service class. Moreover, some
attributes from MongoDB driver to the properties of Shortcut class.

The database provider has been changed. Instead of mLab, now using
MongoDB Atlas, with all IP addresses allowed. The connection now works
flawlessly, and the application has been able to read data successfully.