Implement redirect server in Node.js
Since ASP.NET Core implementation has not been working properly with
my custom domain on Heroku, and Netlify does not support ASP.NET Core
hosting at all, this seems to be the best way: implementation in Node.js
and hosting on Netlify with custom domain, with a free SSL certificate!

Used TypeScript and Express.js, with `mongodb` package for connecting to
the database.